Hotfixfeature/develop

+++ src/Model/CustomValue.php
@@ -13,6 +13,7 @@ use Exceedone\Exment\Enums\ValueType;
 use Exceedone\Exment\Enums\FormActionType;
 use Exceedone\Exment\Enums\ErrorCode;
 use Exceedone\Exment\Enums\JoinedOrgFilterType;
+use Exceedone\Exment\Enums\Permission;
 
 abstract class CustomValue extends ModelBase
 {
@@ -26,6 +27,7 @@ abstract class CustomValue extends ModelBase
     protected $appends = ['label'];
     protected $hidden = ['laravel_admin_escape'];
     protected $keepRevisionOf = ['value'];
+    protected $keepRevisionOfTrigger = ['deleted_at' => 'value'];
 
     /**
      * remove_file_columns.
@@ -84,6 +86,24 @@ abstract class CustomValue extends ModelBase
         // return resuly using cache
         return CustomTable::getEloquent($this->custom_table_name);
     }
+
+    public function getDeletedUserAttribute()
+    {
+        return $this->getUser('deleted_user_id');
+    }
+    public function getDeletedUserValueAttribute()
+    {
+        return $this->getUserValue('deleted_user_id');
+    }
+    public function getDeletedUserTagAttribute()
+    {
+        return $this->getUser('deleted_user_id', true);
+    }
+    public function getDeletedUserAvatarAttribute()
+    {
+        return $this->getUser('deleted_user_id', true, true);
+    }
+
     
     /**
      * Whether this model disable delete
@@ -308,7 +328,7 @@ abstract class CustomValue extends ModelBase
         });
 
         static::deleting(function ($model) {
-            static::setUser($model, ['deleted_user_id']);
+            $model->deleted_user_id = \Exment::user()->base_user_id;
 
             // saved_notify(as update) disable
             $saved_notify = $model->saved_notify;
@@ -320,10 +340,26 @@ abstract class CustomValue extends ModelBase
         });
 
         static::deleted(function ($model) {
+            if($model->isForceDeleting()){
+                return;
+            }
+            
             $model->preSave();
             $model->postDelete();
 
             $model->notify(NotifySavedType::DELETE);
+        });
+
+        static::restored(function ($model) {
+            $model->deleted_user_id = null;
+
+            // saved_notify(as update) disable
+            $saved_notify = $model->saved_notify;
+            $model->saved_notify = false;
+            $model->save();
+            $model->saved_notify = $saved_notify;
+
+            $model->postRestore();
         });
 
         static::addGlobalScope(new CustomValueModelScope);
@@ -649,8 +685,10 @@ abstract class CustomValue extends ModelBase
         // delete role group
         RoleGroupUserOrganization::deleteRoleGroupUserOrganization($this);
 
-        // remove history
-        $this->revisionHistory()->delete();
+        // remove history if hard deleting
+        if($this->isForceDeleting()){
+            $this->revisionHistory()->delete();
+        }
     }
 
     /**
@@ -1235,7 +1273,7 @@ abstract class CustomValue extends ModelBase
         if (($code = $this->custom_table->enableEdit($checkFormAction)) !== true) {
             return $code;
         }
-
+        
         if (!$this->custom_table->hasPermissionEditData($this)) {
             return ErrorCode::PERMISSION_DENY();
         }
@@ -1248,11 +1286,15 @@ abstract class CustomValue extends ModelBase
         if ($this->lockedWorkflow()) {
             return ErrorCode::WORKFLOW_LOCK();
         }
-        
+
         if (!is_null($parent_value = $this->getParentValue()) && ($code = $parent_value->enableEdit($checkFormAction)) !== true) {
             return $code;
         }
 
+        if($this->trashed()){
+            return ErrorCode::ALREADY_DELETED();
+        }
+        
         return true;
     }
 
@@ -1286,6 +1328,45 @@ abstract class CustomValue extends ModelBase
         
         if (!is_null($parent_value = $this->getParentValue()) && ($code = $parent_value->enableDelete($checkFormAction)) !== true) {
             return $code;
+        }
+
+        return true;
+    }
+    
+    /**
+     * User can share this custom value
+     */
+    public function enableShare()
+    {
+        // if system doesn't use role, return false
+        if (!System::permission_available()) {
+            return false;
+        }
+
+        if($this->trashed()){
+            return false;
+        }
+        
+        $custom_table = $this->custom_table;
+
+        // if master, false
+        if (in_array($custom_table->table_name, SystemTableName::SYSTEM_TABLE_NAME_MASTER())) {
+            return false;
+        }
+
+        // if custom table has all_user_editable_flg, return false(not necessary use share)
+        if (boolval($custom_table->getOption('all_user_editable_flg'))) {
+            return false;
+        }
+
+        // if not has edit data, return false
+        if (!$custom_table->hasPermissionEditData($this)) {
+            return false;
+        }
+
+        // if not has share data, return false
+        if (!$custom_table->hasPermission(Permission::CUSTOM_VALUE_SHARE)) {
+            return false;
         }
 
         return true;
